Change download method to correctly load bytes
