+++ sm/plugins/tedge_apt_plugin/src/main.rs
@@ -1,4 +1,3 @@
-use std::path::Path;
 use std::process::{Command, ExitStatus, Stdio};
 use structopt::StructOpt;
